pyproject.toml
[tool.poetry]
name = "eso-tools"
version = "0.1.0"
description = ""
authors = ["Jani Sourander <jani.sourander@kamk.fi>"]
readme = "README.md"
packages = [{include = "eso_tools"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.30.0"
watchdog = "^3.0.0"
pydantic = "^2.5.3"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
